# fix: only `eagerlyLoadType` on introspection requests