--- a/tests/wpunit/AssertValidSchemaTest.php
+++ b/tests/wpunit/AssertValidSchemaTest.php
@@ -27,7 +27,9 @@ class AssertValidSchemaTest extends \Tests\WPGraphQL\TestCase\WPGraphQLTestCase 
 	public function testSchema() {
 		try {
 			$request = new \WPGraphQL\Request();
-			$request->schema->assertValid();
+
+			$schema = WPGraphQL::get_schema();
+			$schema->assertValid();
 
 			// Assert true upon success.
 			$this->assertTrue( true );