--- a/tests/wpunit/CustomPostTypeTest.php
+++ b/tests/wpunit/CustomPostTypeTest.php
@@ -1690,12 +1690,14 @@ class CustomPostTypeTest extends \Tests\WPGraphQL\TestCase\WPGraphQLTestCase {
 			// register the post type with underscore in the graphql_single_name / graphql_plural_name
 			register_post_type( 'with_underscore', $args );
 
-			$request = new \WPGraphQL\Request();
+			new \WPGraphQL\Request();
+
+			$schema = WPGraphQL::get_schema();
 
 			// clean up
 			unregister_post_type( 'with_underscore' );
 
-			$request->schema->assertValid();
+			$schema->assertValid();
 
 			// Assert true upon success.
 			$this->assertTrue( true );
@@ -1730,7 +1732,8 @@ class CustomPostTypeTest extends \Tests\WPGraphQL\TestCase\WPGraphQLTestCase {
 		);
 
 		$request = new \WPGraphQL\Request();
-		$request->schema->assertValid();
+		$schema  = WPGraphQL::get_schema();
+		$schema->assertValid();
 
 		self::assertQuerySuccessful(
 			$actual,
@@ -1758,7 +1761,8 @@ class CustomPostTypeTest extends \Tests\WPGraphQL\TestCase\WPGraphQLTestCase {
 		// assert that the schema is still valid, even though the tax
 		// didn't provide the single/plural name (it will be left out of the schema)
 		$request = new \WPGraphQL\Request();
-		$request->schema->assertValid();
+		$schema = WPGraphQL::get_schema();
+		$schema->assertValid();
 
 		// Cleanup
 		unregister_post_type( 'cpt_no_single_plural' );