--- a/tests/wpunit/CustomPostTypeTest.php
+++ b/tests/wpunit/CustomPostTypeTest.php
@@ -1736,6 +1736,7 @@ class CustomPostTypeTest extends \Tests\WPGraphQL\TestCase\WPGraphQLTestCase {
 		$request = new \WPGraphQL\Request();
 		$schema  = WPGraphQL::get_schema();
 		$schema->assertValid();
+		$this->assertTrue( $schema->hasType( 'NoPlural' ) );
 
 		self::assertQuerySuccessful(
 			$actual,
@@ -1762,9 +1763,14 @@ class CustomPostTypeTest extends \Tests\WPGraphQL\TestCase\WPGraphQLTestCase {
 
 		// assert that the schema is still valid, even though the tax
 		// didn't provide the single/plural name (it will be left out of the schema)
-		$request = new \WPGraphQL\Request();
+		new \WPGraphQL\Request();
 		$schema = WPGraphQL::get_schema();
 		$schema->assertValid();
+
+		$map = array_keys( $schema->getTypeMap() );
+
+		$this->assertTrue( ! in_array( 'BootstrapPost', $map, true ) );
+		$this->assertTrue( ! in_array( 'CptNoSinglePlural', $map, true ) );
 
 		// Cleanup
 		unregister_post_type( 'cpt_no_single_plural' );