What does this implement/fix? Explain your changes.
---------------------------------------------------
Any type that is registered to `eagerlyLoadType` will now only be eagerly loaded for Introspection requests (typically used for tooling such as GraphiQL IDE, Codegen, etc)

For standard execution of queries the eagerly loaded types will not be eagerly loaded which dramatically reduces the number of Object Types that are required to be generated per request.

This introduces: 

- WPGraphQL::set_is_introspection_query
- WPGraphQL::is_introspection_query

Does this close any currently open issues?
------------------------------------------
Closes #3159 


Any other comments?
-------------------
This _might_ be a breaking change if shipped as-is, as the Schema is now generated later in the lifecycle. 

### Performance Tests

I ran some (super basic) performance tests using k6. 

<details>
  <summary>(performance-test.js)</summary>
    
    ```js
    import http from 'k6/http'
    import { check, sleep } from 'k6'
    
    export const options = {
        vus: 10,
        duration: '10s',
    };
    
    export default function() {
        const url = 'http://uri-debugging.local/graphql'
        const query = `
        query {
            posts(first: 10) {
                nodes {
                    __typename
                    title
                    date
                }
            }
        }
        `;
    
    
        // format the query as a query param and append to the url
        // const res = http.get(`${url}?query=${encodeURIComponent(query)}`)
    
        const headers = { 'Content-Type': 'application/json' }
        const res = http.post(url, JSON.stringify({ query: query }), { headers: headers })
        sleep(1)
    
        check(res, {
            'status was 200': r => r.status === 200,
            'transaction time OK (less than 200ms)': r => r.timings.duration < 200
        });
    
    }
    ```
</details>

The test simulates 10 concurrent users executing the following query as a POST request and continues to execute as many times as it can for 10 seconds.

```graphql
    query {
        posts(first: 10) {
            nodes {
                __typename
                title
                date
            }
        }
    }
```

#### BEFORE 

```bash
http_req_duration..............: avg=156.21ms min=52.89ms med=109.72ms max=585.73ms p(90)=331.21ms p(95)=371.64ms
```

#### AFTER

```bash
http_req_duration..............: avg=67.21ms  min=44.3ms  med=56.27ms max=279.04ms p(90)=70.38ms p(95)=164.29ms
```