@justlevine 2024-07-12
Tests passing against this branch for:
- wp-graphql-gravity-forms
- wp-graphql-facetwp
- wp-graphql-headless-login
- wp-graphql-rank-math

Of them, Rank Math is strangely experiencing a significant slowdown on interface queries, but that could also be something with my local machine. I'll try to run some k6/xhprof benchmarks over the weekend and see if I can spot a real discrepancy between this branch and `develop`