@jasonbahl 2024-07-12
@justlevine sounds good! Ya as much as I want to get this out ASAP I want to get it "right" more than get it out fast