@justlevine 2024-07-14
@jasonbahl not sure if its because I'm stuck trying to run these tests on LocalWP + Windows, but while i'm seeing marginal->modest improvements on object queries, queries that rely on interfaces seem to be slower. Would recommend you do some performance comparisons locally as well.

DMing you the test conditions + data I used.